src/main/java/org/bank/service/CustomerService.java
package org.bank.service;

import org.bank.model.Customer;
import org.bank.repository.CustomerRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.Optional;

@Service
public class CustomerService {
    private final CustomerRepository customerRepository;

    @Autowired
    public CustomerService(CustomerRepository customerRepository) {
        this.customerRepository = customerRepository;
    }

    /**
     * Retrieve customer by their unique ID.

     * @param customerId Unique ID of customer to retrieve.
     * @return The customer with the specified ID or null if not found.
     */
    public Customer getCustomerById(Long customerId) {
        Optional<Customer> customer = customerRepository.findById(customerId);
        return customer.orElse(null);
    }
}
